Correct mismatched types in headers

+++ Multiplex/Modules/IDEHeaders/IDEHeaders/DVTKit/DVTSourceBufferProvider-Protocol.h
@@ -11,7 +11,7 @@
 @protocol DVTSourceBufferProvider <NSObject>
 - (unsigned long long)leadingWhitespacePositionsForLine:(unsigned long long)arg1;
 - (struct _NSRange)lineRangeForCharacterRange:(struct _NSRange)arg1;
-- (unsigned long long)length;
+- (NSUInteger)length;
 - (NSString *)string;
 
 @optional
